Changing magics creation so that they don't get created if not in ipython (#332)
Adding import of magics to nbinit and removing from __init__ and Pivot class.
Updating docs (including some auto-gen'd)
