+++ docs/source/api/msticpy.data.drivers.rst
@@ -4,6 +4,14 @@ msticpy.data.drivers package
 Submodules
 ----------
 
+msticpy.data.drivers.cybereason\_driver module
+----------------------------------------------
+
+.. automodule:: msticpy.data.drivers.cybereason_driver
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 msticpy.data.drivers.driver\_base module
 ----------------------------------------
