+++ msticpy/datamodel/entities/entity.py
@@ -86,7 +86,7 @@ class Entity(ABC, Node):
         self.TimeGenerated = datetime.utcnow()
         self.Type = self._get_entity_type_name(type(self))
         # If we have an unknown entity see if we a type passed in
-        if self.Type == "unknownentity" and "Type" in kwargs:
+        if self.Type == "unknown" and "Type" in kwargs:
             self.Type = kwargs["Type"]
         # Make sure Type is in the class schema dictionary
         self._entity_schema["Type"] = None
@@ -521,8 +521,8 @@ class Entity(ABC, Node):
                 )
             )
         except StopIteration:
-            name = None
-        return name or "unknown"
+            name = "unknown"
+        return name
 
     @property
     def node_properties(self) -> Dict[str, Any]:
