+++ msticpy/nbtools/nbinit.py
@@ -154,6 +154,7 @@ _MP_IMPORTS = [
     dict(pkg="msticpy.common.wsconfig", tgt="WorkspaceConfig"),
     dict(pkg="msticpy.datamodel.pivot", tgt="Pivot"),
     dict(pkg="msticpy.datamodel", tgt="entities"),
+    dict(pkg="msticpy.nbtools", tgt="nbmagics"),
     dict(pkg="msticpy.vis", tgt="mp_pandas_plot"),
 ]
 _MP_IMPORT_ALL = [
